refactor: Streamline serialization methods and enhance type handling in ReflectionSerializerBase
