+++ Assets/root/Runtime/Reflection/Custom/RS_UnityEngineObject.cs
@@ -2,7 +2,6 @@
 using System;
 using System.Reflection;
 using System.Text.Json;
-using com.IvanMurzak.Unity.MCP.Common;
 using com.IvanMurzak.Unity.MCP.Common.Data.Utils;
 using UnityEngine;
 
@@ -25,8 +24,8 @@ namespace com.IvanMurzak.Unity.MCP.Utils
                     {
                         name = name,
                         type = type.FullName,
-                        fields = Serializer.SerializeFields(obj, flags),
-                        properties = Serializer.SerializeProperties(obj, flags)
+                        fields = SerializeFields(obj, flags),
+                        properties = SerializeProperties(obj, flags)
                     }.SetValue(new InstanceID(unityObject.GetInstanceID()));
                 }
                 else
@@ -43,16 +42,14 @@ namespace com.IvanMurzak.Unity.MCP.Utils
         {
             return true;
         }
-        protected override bool SetFieldValue(ref object obj, Type type, FieldInfo fieldInfo, JsonElement? value)
+        protected override bool SetFieldValue(ref object obj, Type type, FieldInfo fieldInfo, object? value)
         {
-            var parsedValue = JsonUtils.Deserialize(value.Value.GetRawText(), type);
-            fieldInfo.SetValue(obj, parsedValue);
+            fieldInfo.SetValue(obj, value);
             return true;
         }
-        protected override bool SetPropertyValue(ref object obj, Type type, PropertyInfo propertyInfo, JsonElement? value)
+        protected override bool SetPropertyValue(ref object obj, Type type, PropertyInfo propertyInfo, object? value)
         {
-            var parsedValue = JsonUtils.Deserialize(value.Value.GetRawText(), type);
-            propertyInfo.SetValue(obj, parsedValue);
+            propertyInfo.SetValue(obj, value);
             return true;
         }
     }
