+++ Assets/root/Runtime/Reflection/Custom/ReflectionSerializerBase.Serialize.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ea73303894b7a40488dcd28abd4adb67
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
