+++ Assets/root/Runtime/Reflection/Serializer.API.cs
@@ -1,12 +1,9 @@
 #pragma warning disable CS8632 // The annotation for nullable reference types should only be used in code within a '#nullable' annotations context.
 using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Reflection;
 using System.Text;
 using com.IvanMurzak.Unity.MCP.Common.Data.Utils;
 using com.IvanMurzak.Unity.MCP.Common.Utils;
-using UnityEngine;
 
 namespace com.IvanMurzak.Unity.MCP.Utils
 {
@@ -32,46 +29,7 @@ namespace com.IvanMurzak.Unity.MCP.Utils
             }
             throw new ArgumentException($"[Error] Type '{type?.FullName}' not supported for serialization.");
         }
-        public static List<SerializedMember> SerializeFields(object obj, BindingFlags flags)
-        {
-            var serialized = default(List<SerializedMember>);
-            var objType = obj.GetType();
-
-            foreach (var field in objType.GetFields(flags)
-                .Where(field => field.GetCustomAttribute<ObsoleteAttribute>() == null)
-                .Where(field => field.IsPublic || field.IsPrivate && field.GetCustomAttribute<SerializeField>() != null))
-            {
-                var value = field.GetValue(obj);
-                var fieldType = field.FieldType;
-
-                serialized ??= new();
-                serialized.Add(Serialize(value, fieldType, name: field.Name, recursive: false, flags: flags));
-            }
-            return serialized;
-        }
-
-        public static List<SerializedMember> SerializeProperties(object obj, BindingFlags flags)
-        {
-            var serialized = default(List<SerializedMember>);
-            var objType = obj.GetType();
-
-            foreach (var prop in objType.GetProperties(flags)
-                .Where(prop => prop.GetCustomAttribute<ObsoleteAttribute>() == null)
-                .Where(prop => prop.CanRead))
-            {
-                try
-                {
-                    var value = prop.GetValue(obj);
-                    var propType = prop.PropertyType;
-
-                    serialized ??= new();
-                    serialized.Add(Serialize(value, propType, name: prop.Name, recursive: false, flags: flags));
-                }
-                catch { /* skip inaccessible properties */ }
-            }
-            return serialized;
-        }
-        public static StringBuilder Populate<T>(ref T obj, SerializedMember data, StringBuilder stringBuilder = null, int depth = 0,
+        public static StringBuilder Populate(ref object obj, SerializedMember data, StringBuilder stringBuilder = null, int depth = 0,
             BindingFlags flags = BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Instance)
         {
             stringBuilder ??= new StringBuilder();
@@ -86,17 +44,15 @@ namespace com.IvanMurzak.Unity.MCP.Utils
             if (obj == null)
                 return stringBuilder.AppendLine(new string(' ', depth) + Error.TargetObjectIsNull());
 
-            var castedObj = TypeUtils.CastTo(obj, data.type, out var error);
+            TypeUtils.CastTo(obj, data.type, out var error);
             if (error != null)
                 return stringBuilder.AppendLine(new string(' ', depth) + error);
 
-            if (!type.IsAssignableFrom(castedObj.GetType()))
+            if (!type.IsAssignableFrom(obj.GetType()))
                 return stringBuilder.AppendLine(new string(' ', depth) + Error.TypeMismatch(data.type, obj.GetType().FullName));
 
-            var castedObject = (object)obj;
-
             foreach (var populators in Registry.BuildPopulatorsChain(type))
-                populators.Populate(ref castedObject, data, stringBuilder: stringBuilder, flags: flags);
+                populators.Populate(ref obj, data, stringBuilder: stringBuilder, flags: flags);
 
             return stringBuilder;
         }
